加了更改offsett添加陀螺仪pid参数没改好，0.04飘

+++ SHOOT/Chassic3.0/Src/IOTask.h
@@ -43,9 +43,9 @@ typedef __packed struct
 
 typedef __packed struct
 {
-    int16_t     GyroXOffset;
-    int16_t     GyroYOffset;
-    int16_t     GyroZOffset;
+    float     GyroXOffset;
+    float     GyroYOffset;
+    float     GyroZOffset;
     uint8_t     GyroCaliFlag;
 }GyroCaliStruct_t;
